rename components from index.jsx
