+++ components/CoinDetailModal/CoinDetailModal.jsx
