+++ components/WatchListSection/WatchlistSection.jsx
